html/semantics/forms/the-selectmenu-element/selectmenu-tab-navigation.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
focused_element_id = null;

function OnFocus(event) {
  focused_element_id = event.target.id;
}
</script>

<input id="input1" onfocus="OnFocus(event)">
<selectmenu id="selectmenu" onfocus="OnFocus(event)">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>
<input id="input3" onfocus="OnFocus(event)">

<script>
promise_test(async () => {
  const TAB_KEY = "\uE004";

  const input1 = document.getElementById("input1");
  const selectmenu = document.getElementById("selectmenu");
  const input3 = document.getElementById("input3");

  input1.focus();
  assert_equals(focused_element_id, "input1", "input1 should be focused");

  await test_driver.send_keys(input1, TAB_KEY);
  assert_equals(focused_element_id, "selectmenu", "selectmenu should be focused");

  await test_driver.send_keys(selectmenu, TAB_KEY);
  assert_equals(focused_element_id, "input3", "input3 should be focused");
}, "Check that <selectmenu> occupies just one slot in tab navigation.");
</script>
